TCP M9 a medias

+++ eclipse-workspaces/M06/UF2/hibernate/src/main/java/services/DepartamentosService.java
@@ -7,7 +7,7 @@ import pojos.Departaments;
 public interface DepartamentosService {
     public List<Departaments> getAll();
 
-    public Departaments selectOne(int deptNo);
+    public Departaments selectOne(byte deptNo);
 
     public Boolean createOne(Departaments departamento);
